(chore) move permissions and journal into src folder
