+++ packages/Backend/index.js
@@ -3,5 +3,5 @@ import { config } from 'firebase-functions';
 
 admin.initializeApp(config().firebase);
 
-export * from './permissions';
-export { default as submit } from './journal/submit';
+export * from './src/permissions';
+export { default as handleSubmit } from './src/journal/handleSubmit';
